+++ server/services/chat-gpt.service.js
@@ -1,9 +1,6 @@
 "use strict";
-const https = require('https');
-const fs = require('fs');
-const crypto =  require("crypto");
 const { OpenAI } = require("openai");
-const mime = require('mime-types'); //used to detect file's mime type
+const utils = require('../utils')
 
 module.exports = ({ strapi }) => ({
   async getResponsefromChatGpt(ctx) {
@@ -84,7 +81,7 @@ module.exports = ({ strapi }) => ({
         size,
       });
 
-      const savedFile = await saveFile(data.data[0].url)
+      const savedFile = await utils.saveFile(data.data[0].url)
 
       return { response: [savedFile, data.data[0].url]};
     } catch (error) {
@@ -108,7 +105,7 @@ module.exports = ({ strapi }) => ({
         name: "strapi-supergpt",
       });
 
-      return pluginStore.get({ key: "strapiChatGPTConfig" });
+      return pluginStore.get({ key: "superGPTConfig" });
     } catch (error) {
       strapi.log.error(error.message);
       return {
@@ -139,7 +136,7 @@ module.exports = ({ strapi }) => ({
       });
 
       return pluginStore.set({
-        key: "strapiChatGPTConfig",
+        key: "superGPTConfig",
         value: data,
       });
     } catch (error) {
@@ -152,44 +149,3 @@ module.exports = ({ strapi }) => ({
   },
 });
 
-async function saveFile(url) {
-  return new Promise((resolve, reject) => {
-    strapi.log.info(`Saving the picture ${url}`);
-    const fileName = crypto.randomUUID();
-    const rootDir = process.cwd();
-    const filePath = `${rootDir}/public/uploads/${fileName}.png`;
-    const file = fs.createWriteStream(filePath);
-    const request = https.get(url, function(response) {
-      response.pipe(file);
-
-      file.on("finish", async () => {
-        file.close();
-        strapi.log.info("Download Completed");
-
-        // Ensure the file is closed before accessing its stats
-        fs.stat(filePath, async (err, stats) => {
-          if (err) {
-            reject(strapi.log.error('Error getting file stats'));
-          }
-
-          // Now that the download is complete and file is closed, upload it
-          const upload = await strapi.plugins.upload.services.upload.upload({
-            data: {}, // mandatory declare the data (can be empty)
-            files: {
-              path: filePath,
-              name: `${fileName}.png`,
-              type: mime.lookup(filePath), // mime type of the file
-              size: stats.size,
-            },
-          });
-          strapi.log.info('Upload Completed');
-          resolve(upload[0].formats['medium'].url)
-        });
-      });
-    });
-
-    request.on('error', (e) => {
-      reject(strapi.log.error(`Request error: ${e.message}`));
-    });
-  })
-}
